add a simple kf for estimating gyro bias

+++ sv/llol/lidar.cpp
@@ -55,12 +55,12 @@ cv::Point3f LidarModel::Backward(int r, int c, float rg) const {
 
 int LidarModel::ToRow(float z, float r) const {
   const float elev = std::asin(z / r);
-  return (elev_max - elev) / elev_delta + 0.5F;
+  return static_cast<int>((elev_max - elev) / elev_delta + 0.5F);
 }
 
 int LidarModel::ToCol(float x, float y) const {
   const float azim = std::atan2(y, -x) + kPiF;
-  return azim / azim_delta + 0.5F;
+  return static_cast<int>(azim / azim_delta + 0.5F);
 }
 
 std::string LidarModel::Repr() const {
